feat(contract-distribution): Exclude old-code from witness for deploy-contract and delete-account actions (#12416)
In contract-distribution feature
(`ExcludeContractCodeFromStateWitness`), we record contract accesses for
function calls and we do not include the contract code in the state
witness; instead we distribute these contracts separately to the
validators.

There are also two other actions (`DeployContract` and `DeleteAccount`)
that are seen less frequently but also reads code from the storage and
contribute to state witness. Differently from function call, these
actions do not run the code but only capture the code length in order to
update the account's storage usage. Thus, there is no need to read and
send the actual code to the validators in the witness. To address this,
we change the implementation of these actions to read the code length
without reading the actual code.

**NOTE:** The changes are guarded by the new protocol feature
`ExcludeExistingCodeFromWitnessForCodeLen` (added to Nightly), so it
will be enabled after `2.4.0`.

The main files to review are `actions.rs` and `update.rs`.
Related changes in this PR:

- As a refactor, move `get_code` and `set_code` function from
`core\store\src\lib.rs` to `TrieUpdate` as member methods (and the new
method mentioned below). Note that, this will potentially be leveraged
to change the code further to use `ContractStorage` instead of the trie
and `prospective` changes. Note that `get_code` took `TrieAccess` before
but all calls are done via `TrieUpdate`.

- Add a new method to `TrieUpdate` called `get_code_len`, which calls
`get_ref` to access the reference and capture the length of the value.
Note that `get_ref` also checks `prospective` and `committed` lists to
see the intermediate code changes.

- Update the `DeployContract` and `DeleteAccount` actions to use this
new method instead of reading the whole code.

- Also use this method in `TrieViewer` (since it also uses `get_code` to
read the code length.

- Testloop test for this change are being added in separate PR in
https://github.com/near/nearcore/pull/12428.