[cleanup] Remove no_cache feature (#12459)
In future PRs we plan to get rid of the caches in ChainStore as we don't
believe them to be too useful and is a complication to the current
codebase.

As a pre-step, this PR gets rid of the `no_cache` feature. This was
being used in ChainStore TrieStore, and EpochManager.

Here are the associated caches that are affected
- [removed] ChainStore: We are effectively removing the cache from chain
store
- [retained] Epoch Manager: While we are removing the no_cache feature,
we maintain the default cache sizes for EpochManager caches
- [retained] Trie Cache: While we are removing the no_cache feature, we
maintain the default cache sizes for Trie caches.

## Effect on Parameter Estimator

While the param estimator hasn't been run for a long time since the
introduction of memtries, it used to rely on the `no_cache` feature to
effectively test the worst case performance with the cache removed from
Trie.

With the introduction on memtrie, this check is now redundant as we no
longer use the Trie Store Cache after the introduction of memtrie.

However, as of today, some nodes like RPC nodes (which track all shards)
do not use memtries and this might be relevant for them. That said, it's
probably not worth it to block this change on potential future use of
`no_cache` for RPC nodes where we are moving towards a memtrie world.

## Native transfer workloads

This change is backed up by Moritz's tests on native transfer workload
with 10k accounts and unlimited configs.
- cargo build -p neard --release @ 2.3.0: between 3000 and 3100 TPS
- cargo build -p neard --release --features no_cache @ 2.3.0: between
3000 and 3100 TPS
- cargo build -p neard --release @ 2.3.0 + this PR patch: between 3050
and 3150 TPS

This shows that the caches don't have a significant impact on the
workloads.