state-dumper: Don't store state parts in the database (#12486)
When the state dumper genrates state parts to upload to GCS, it stores
them in the StateParts column before doing so. But it never reads these
parts after storing them. Storing these parts is not a free operation,
because on a forknet run with recent mainnet state, just the storage of
the part took on the order of 50-100 millis. Over several thousand
parts, that's a nontrivial slowdown for no good reason.