fix(state-dumper): dump state for post-resharding shards (#12491)
Currently, the state dumper starts one thread for each ShardId in the
current epoch, and each of those is responsible for dumping headers and
parts just for that ShardId. But after a resharding, there's no thread
that's aware that it should dump state for any of the new ShardIds. Here
we fix it by iterating not just over the ShardIds in the current epoch
when we start the threads, but also over any future ShardIds that might
belong to post-protocol upgrade epochs.

This is not great because we're starting threads that won't be doing
anything useful (but still doing work in a loop which in tests can be
nontrivial since we set the "iteration_delay" config value to 100ms) for
quite some time, and we don't stop old threads after the shard ID they
correspond to is no longer a valid shard ID in the current epoch. But
it's not horrible and this is an easy first fix.