chore: update crates (#12482)
This PR updates a number of dependency versions, mostly focusing on
reducing the scope at which we depend on `syn 1.0`. This ultimately is a
net-zero in the crate number as the dependency on syn ^1 couldn't be
removed quite yet, and whatever wins I was able to eke out went back
into duplication for `wasm-tools`.

`wasm-tools` updates in particular help us to get better coverage in our
fuzzers with e.g. `wasm-smith` generating more involved test samples
than it knew how to do before. Related PR
https://github.com/near/finite-wasm/pull/78 helps with reducing the
additional duplication here.

In updates herein I took care to not touch crates that might affect the
protocol behaviour, only updating those that seemed particularly safe to
update. However there is one crate that I will update in a follow-up
that's going to be ever-so-slightly more risky: rkyv.