feat(resharding) - load memtries pre-resharding (#12484)
The shard trie must be loaded into memory before it can be split in
resharding. As of today some nodes do not enable memtrie so it is
necessary to load the trie into memory on demand before resharding. This
is the implementation of option 2, loading the shards into memory on
startup, for all shards that will be resharding in this client / binary.