feat(mirror): Clear state parts gcp bucket on reset (#12502)
The bucket is created along with the state dumper node. If the state
dumper node is present we want to clear the bucket every time we reset
the network to a previous state.
This can be on `hard-reset`, `new-test` or when we run `reset` to
restore from a backup.

The command will be executed on the dumper node to make sure that it has
the edit access to the bucket.