[trie] Introduce GenericTrieNode as replacement of GenericUpdatedTrieNode (#12496)
This PR introduces the GenericTrieNode which is a generalization on top
of GenericUpdatedTrieNode where the nodes need not be an update type
node.

This is useful in scenarios where we are not dealing with just updates,
for example generalizing iterators over tries etc.

Along with this change this PR renames `GenericUpdatedNodeId` to
`UpdatedNodeId` which is more appropriate as for both memtries and store
tries we have UpdatedNodeId as usize.