fix(resharding): Skip old ShardUIds in get_new_flat_storage_head() (#12505)
This function finds the chunk in the last final block by looking for the
one with the same shard ID as one of the shards in the current block's
epoch. But if there is a skipped chunk in a child shard in the first
block after a resharding, then the old chunk in the child's shard_index
will be the parent's chunk from the previous epoch. Then this shard ID
comparison doesn't work and we get an error. Fix it by comparing the
current ShardUID to the ShardUID of the chunk, and returning None if
they don't match.