fix(tests) - fix test_protocol_upgrade always upgrading to the latest version (#12514)
`test_protocol_upgrade` (added in
https://github.com/near/nearcore/pull/12253) is supposed to upgrade from
`old_protocol` to `new_protocol`, but it actually did an upgrade from
`old_protocol` to `PROTOCOL_VERSION` (latest protocol). It just so
happened that all uses of the test were with `new_protocol` equal to
`PROTOCOL_VERSION` and the bug wasn't noticed when the test was written.

I did some debugging and it looks like the problem is that clients are
voting for the latest protocol instead of the one specified in
`new_protocol`, and switch to the latest one instead of the one that
they were supposed to.

The protocol version that nodes vote for is controlled by
`PROTOCOL_UPGRADE_SCHEDULE`, which is a global constant. It being a
global constant makes it awkward to overwrite in tests.

I dealt with the problem by adding an `upgrade_schedule` configuration
parameter to `Client`, which allows to set the voting schedule that this
individual client will use. With this in place, we can override the
voting schedule for clients in `test_protocol_upgrade` and upgrade to
the correct protocol version.

PR is meant to be reviewed commit-by-commit, I added some more comments
in commit descriptions.