test(resharding) - un-ignore buffered receipts test (#12529)
Un-ignored the buffered receipts test and added an extra check that the
buffered receipts are present after resharding.

There is a bit of refactoring here, mostly moving most of the logic of
`check_receipts_presence_at_resharding_block` to a new method
`check_receipts_at_block` that is now used for both before and after
resharding checks.