fix: Remove annotations for not sending fields if zero or empty (#12525)
This is causing debug page UI to render NaN, as we do not send some
values.

We could also fix the UI code but still sending zero/empty values is
consistent with block/chunk production fields and should not hurt to
send 0/empty.

Testing: Tested with mainnet RPC node with the change:
http://debug.nearone.org/34.90.80.187/epoch_info/validators