feat(archival-store): Add ArchivalStoreConfig to configure external storage locations (#12551)
Currently ColdDB is used for archival storage and it is configured by
the `Config.cold_store` field (which contains a `StoreConfig`).

In order to prepare for introducing other storage alternatives, we
introduce a new config called `ArchivalStoreConfig`. It defaults to
`ColdDB`. We also change `NodeStorage::opener` to accept
`Option<ArchivalConfig>` where `ArchivalConfig` contains references to a
couple of archival-node related configurations, since the archival
storage will be configured by multiple config objects.