feat(resharding) - buffered receipts handling (#12538)
Implemented handling for buffered receipts and un-ignored the tests. Now
both of the following cases should work:
* buffered receipts in the parent are inherited by the left child
* buffered receipts in other shards, aimed for the parent shard, are now
reassigned to the correct child shard

Some minor refactoring here and there:
* Added shard layout to the epoch info provider
* Less verbose Debug for ShardId

This PR does not touch on the congestion info that needs to be set
correctly. That'll be done separately.

The `debug_check_unchanged` came in very handy in detecting various
issues.
* Forwarding from the parent twice, one for each child. 
* Forwarding from the same shard twice because every shard is its own
parent.