mirror: don't require an archival dir for the traffic generator (#12559)
The reason this was done originally was to make sure we apply all chunks
so that in `on_target_block()` we can look through all the outcomes and
keep track of how many of our txs made it on chain, and so that we can
figure out what unstake actions need to be sent to undo stakes that came
from staking pool contract calls. The first one is just best effort
anyway, and the second one won't even be that relevant in the case where
the traffic generator is offline for several epochs and state syncs,
since by that time it will be too late to unstake them. So just remove
this check and don't make the traffic generator an archival node