test(resharding) - Test resharding with both V1 and V2 base shard layout (#12576)
* Added base_shard_layout_version to the resharding test loop params. 
* Refactored the buffered receipts tests to have one test for each base
shard layout
* Refactored the load memtrie tests to have one test for each base shard
layout
* ~Removed what seems to be an old copy of some tests - we have tests
for both `buffered` and `outgoing` receipts that look like old and new
versions of the same test. Please let me know if that's not the case and
those were actually testing different things.~

Let me know if any other tests deserve to cover both cases. For now the
default is set to V2 because that's what we'll use in the long term.

I'm curious if there is any rust library for automatically creating
tests for various test parameters combinations. Thoughts on whether that
would be useful?