Fix persistence of the Receipts column if we don't track all shards. (#12613)
Detailed discussion here
https://near.zulipchat.com/#narrow/channel/295558-core/topic/DBCol.3A.3AReceipts.20.2B.20single.20shard.20tracking

The way we choose the receipts to write to DBCol::Receipts is incorrect
if the node does not track all shards. The prev_outgoing_receipts of a
chunk do not represent the receipts executed by the chunk, but rather
the outgoing receipts from the execution of the previous chunk, which
target arbitrary shards. In other words, if a node tracks only some
shards, a receipt may be executed by the node but not actually in the
Receipts column.

To fix this, we change the DBCol::Receipts column to be an exact
reindexing of the DBCol::IncomingReceipts column, which accurately
reflects the receipts that this node would actually execute.

Even though this changes the semantics of the column, there is no DB
migration, because:
* For RPC or archival nodes or any node that is not a validator,
currently they all track all shards, meaning that the IncomingReceipts
column for the same block is the same as the aggregation of the
prev_outgoing_receipts of all new chunks in this block, so the data in
the DBCol::Receipts (including refcounts) would be identical before and
after this PR.
* For validator nodes, they have no use of the DBCol::Receipts column.
Even though this change would introduce some inconsistency on that
column (GC deleting some receipts that didn't exist before, and failing
to clean up some receipts that existed before), the column is not used
so it doesn't matter. The space wasted from failing to clean up some of
the receipts is very limited compared to how much space could be saved
later via Epoch Sync. So it's not worth doing a migration.

The store_validator has been augmented to properly check for the
equivalence of DBCol::Receipts against DBCol::IncomingReceipts. (We used
to only check for the refcount of whatever receipts existed in
DBCol::Receipts, meaning that the store validator would not have
complained if we persisted no receipts at all in that column).

One additional very subtle fix: for state sync, in set_state_finalize,
we only persist the incoming receipts relevant to the chunk we executed,
rather than the whole incoming_receipts_proofs, because the latter can
actually include receipts *later* than the chunk being executed, causing
double-counting of the sync_hash block's incoming receipts.