fix(resharding): balance checker for ignored delayed receipts (#12619)
Tinkering with resharding testloop in another PR, I got some errors when
creating / removing an account.
One of these errors is balance checker failing.

For delayed receipt test, we use `call_burn_gas_contract` that does not
attach deposit, so balance checker did not fail.
But if a delayed receipt has deposit and it is skipped (because it
belongs to the sibling shard), then balance checker fails,
because it calculated total deposit independently of what delayed
receipts were actually processed.

---------

Co-authored-by: Andrea <Trisfald@users.noreply.github.com>