feat(resharding): resharding archival node (#12578)
Mapping from child to parent shard is stored in
`DBCol::StateShardUIdMapping`, which is marked as a cold column.
By default, it would be copied from hot to cold storage at each block.
In this PR, we do add `is_last_block_in_epoch` parameter to
`update_cold_db()` so that it is only copied once per epoch.

For that, a slight refactor of `cold_store_copy()` was needed.
If it is possible to have skips at epoch boundary, then
`cold_store_copy()` had a bug, where previous epoch shard layout could
be used to copy data for a block from the new shard layout.

Added archival node and rpc node to reshardingv3 testloop.
In testloop, create temporary account before resharding and remove it
after resharding. After gc period, the account is available at the
archival node and it is not available at rpc node.
However, it works without actually testing the `cold_store_copy` path,
because neither `cold_store_copy` nor garbage collection is called for
archival node in testloop. So it needs fixing testloop and a follow-up
testing.