fix(snapshots): store BlockInfo in Resharding(CatchingUp) status (#12651)
https://github.com/near/nearcore/pull/12589 made a change required for
making snapshots of child shards that are catching up after resharding,
where if we want to take a state snapshot of a child shard's flat
storage whose status is `Resharding(CatchingUp)`, then we set it to
`Ready` in the state snapshot. To do this, we need to be able to figure
out what the correct `BlockInfo` is just from the hash, so we read the
info from the block headers column. But this column is not kept in state
snapshots (which was previously overlooked in test loop since everything
was copied). So fix it by storing a `BlockInfo` in the
`Resharding(CatchingUp)` status so we don't need to look these up
anymore