test: update config adjustments for `benchmarknet` (#12659)
Previously, for benchmarknet some fields of `CongestionControlConfig`
and `WitnessConfig` had been modified. As these structs evolved, the
modifications became outdated.

This PR uses `test_disabled()` to generate entire
`CongestionControlConfig` and `WitnessConfig` structs with all limits
disabled. With this approach, fewer updates to benchmarknet adjustments
should be required in the future.

Note that this config is effective only when `chain_id: "benchmarknet"`,
which should happen only in tests or benchmarks.