fix(resharding): Increase epoch length for shard shuffling (#12674)
Some resharding tests were failing with single shard tracking enabled
because of epoch length was sligthly too short.
Under the hood, if we are close to epoch length (4 blocks or less), we
forward transactions to chunk producer for the next epoch too. But if
the epoch length is 6, it means the chunk producer has only 2 blocks to
catchup before it is supposed to start accepting transactions for the
new shard. In such case, `TX_CHECK_BLOCKS_AFTER_RESHARDING` needs also
be increased.

**Changes:**
This PR sets `INCREASED_EPOCH_LENGTH: 8` for these tests.
For remaining tests, I keep the old `DEFAULT_EPOCH_LENGTH: 6`, to test
potentially more corner cases, and some tests seem to be designed with
this specific epoch length in mind.
`TX_CHECK_BLOCKS_AFTER_RESHARDING ` is increased to almost max
(`epoch_length - 2`).
Loop actions are slightly refactored, introducing `LoopAction` to make
sure these actions actually succeed before the test is over. It happened
to me several times they did not, if some parameter was changed.