fix(resharding): MissingTrieValue in tests (#12677)
Partially applies the patch: https://github.com/near/nearcore/pull/12597

It was failing for delayed receipt tests with
`MissingTrieValue(TrieStorage)` inside
`check_state_shard_uid_mapping_after_resharding`.
The reason is that trie access interface returns `None` for a key with
an empty value (which is the case for delayed receipts and negative
refcounts).
That was not caught up before the patch, because the
`check_state_shard_uid_mapping_after_resharding` was called after
resharding, but not in the end of the test (after gc kicked in).

**Changes**
This PR modifies the `check_state_shard_uid_mapping_after_resharding` to
not check entries with negative refcount.
The test is refactored so that the main loop does not finish an epoch
after resharding, but keeps running until gc kicks in.
For that, the code responsible for testing temporary account is moved to
a loop action,
and all loop actions are moved to a separate file for readability.

**Note**
`test_resharding_v3_yield_resume` is still failing with the original
patch: https://github.com/near/nearcore/pull/12597