test(resharding): base resharding scenario state cleanup (#12678)
A test for cleanup after resharding.
Adds an extra client without a role, that initially tracks the parent
shard, then a child shard, and then an unrelated shard.
After a few epochs, only the last tracked shard should reside in State,
while currently the test fails because it has entries from the first
tracked shard too.

**Changes**
- Fix a bug in test infra: pass `gc_num_epochs_to_keep` to the runtime.
- Add `tracked_shard_schedule` option to resharding testloop, in order
to test the cleanup.
- Add the cleanup test itself (ignored for now).